Move BubbleContainer animation into AnimationBubble
